Furthur simplify interfaces file
